DigitalOcen cli

+++ CloudCli/DigitalOcean/scripts/firewall-delete.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
 
 
-ID=$(firewall-id "fr-$1-in-allow-ssh")
+ID=$(firewall-id "$1")
+
+
+if [ -z "$ID" ]; then
+  echo "Firewall $1 doesn't exist"
+  exit 1
+fi
 
 doctl compute firewall delete $ID
